requirements.txt
aerich==0.7.2
aiofiles==23.2.1
aioredis==2.0.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
asyncpg==0.29.0
Babel==2.14.0
bcrypt==4.1.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
dictdiffer==0.9.0
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.1.1
fastadmin==0.1.41
fastapi==0.110.0
fastapi-admin==1.0.4
h11==0.14.0
httptools==0.6.1
idna==3.6
iso8601==1.1.0
isort==5.13.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pendulum==3.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.15
pydantic_core==2.16.3
PyJWT==2.8.0
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
time-machine==2.14.1
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
